Split out main page and added FAQ page
