+++ src/App.tsx
@@ -1,92 +1,32 @@
-import React, { useEffect, useState } from 'react';
+import React from 'react';
 import './App.css';
-import FaqCarousel from './components/FaqCarousel';
 import Navigation from './components/Navigation';
-import video_camera from './images/bg_video_camera.webp';
-import bg_tools from './images/bg_tools.webp';
-import bg_filming from './images/bg_filming.webp';
-import bg_community from './images/bg_community.webp';
-import IconNavigator from './components/IconNavigator';
-import Section from './components/Section';
-import SectionStyle from './enums/SectionStyle';
-import ShopSection from './components/ShopSection';
-import { FirebaseContext } from './firebase';
-import CardDetails from './models/CardDetails';
-import CollectionType from './enums/CollectionType';
-import ExamplesPage from './components/ExamplesPage';
+import DynamicContentPage from './components/DynamicContentPage';
+import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
+import MainPage from './components/MainPage';
+import bg_question_mark from './images/bg_question_mark.webp';
 
 const App: React.FC = () => {
-  const firebaseContext = React.useContext(FirebaseContext);
-  const [communities, setCommunities] = useState<CardDetails[]>([]);
-  const [resources, setResources] = useState<CardDetails[]>([]);
-  const [creators, setCreators] = useState<CardDetails[]>([]);
-  useEffect(() => {
-    firebaseContext?.subscribeToCollection(
-      CollectionType.Communities,
-      (results) => {
-        setCommunities(results);
-      },
-      (error) => {
-        console.log(error);
-      },
-    );
-    firebaseContext?.subscribeToCollection(
-      CollectionType.Resources,
-      (results) => {
-        setResources(results);
-      },
-      (error) => {
-        console.log(error);
-      },
-    );
-    firebaseContext?.subscribeToCollection(
-      CollectionType.Creators,
-      (results) => {
-        setCreators(results);
-      },
-      (error) => {
-        console.log(error);
-      },
-    );
-  }, [firebaseContext]);
-
   return (
-    <div className="App">
-      <Navigation />
-      <div className="container-fluid p-0 mt-5">
-        <a id="faq">
-          <FaqCarousel />
-        </a>
-        <a id="examples">
-          <Section backgroundSource={video_camera} sectionStyle={SectionStyle.Light}>
-            <ExamplesPage />
-          </Section>
-        </a>
-        <a id="resources">
-          <Section heading="Resources" backgroundSource={bg_tools}>
-            <IconNavigator heading={undefined} cardDetails={resources} />
-          </Section>
-        </a>
-        <a id="communities">
-          <Section heading="Communities" backgroundSource={bg_community} sectionStyle={SectionStyle.Light}>
-            <IconNavigator
-              heading={undefined}
-              cardDetails={communities}
-              sectionStyle={SectionStyle.Light}
-              roundCards={true}
-            />
-          </Section>
-        </a>
-        <a id="creators">
-          <Section heading="Content Creators" backgroundSource={bg_filming}>
-            <IconNavigator heading={undefined} cardDetails={creators} roundCards={true} />
-          </Section>
-        </a>
-        <a id="shop">
-          <ShopSection />
-        </a>
+    <Router>
+      <div className="App">
+        <Navigation />
+        <div className="mt-5">
+          <Switch>
+            <Route path="/faq">
+              <DynamicContentPage
+                pageName="faq"
+                title="Frequently Asked Questions"
+                backgroundSource={bg_question_mark}
+              />
+            </Route>
+            <Route path="/">
+              <MainPage />
+            </Route>
+          </Switch>
+        </div>
       </div>
-    </div>
+    </Router>
   );
 };
